# src/lib/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workers Webhooks API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Bearer token authentication"
      },
      "apiKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "apiKeyCookie",
        "description": "API Key authentication via cookie"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "description": "Error message"
                }
              }
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyCookie": []
    }
  ],
...
  "paths": {
...
    "/send": {
      "post": {
        "summary": "Send Event",
        "description": "Send an event to the specified endpoints.\n\n Events can be sent to endpoints and/or endpoint groups. \n In order to send an event to an endpoint group, the eventType must be specified, then the event will be dispatched to all endpoints in the group that are subscribed to the eventType.\n\n API key requires the `webhooks:send` permission to access this endpoint.\n",
        "tags": [
          "Send"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "payload",
                  "destinations"
                ],
                "properties": {
                  "destinations": {
                    "description": "List of endpoints or endpoint groups to send the event to",
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "ep_a1b2_abcd1234",
                      "grp_a1b2_efgh5678"
                    ]
                  },
                  "eventType": {
                    "description": "Event type to send the event to. Required if sending to an endpoint group.",
                    "type": "string",
                    "example": "user.created"
                  },
                  "payload": {
                    "description": "Payload to send with the event",
                    "type": "object",
                    "example": {
                      "userId": "abcd1234"
                    }
                  },
                  "eventId": {
                    "description": "Optional unique event ID to track the event",
                    "type": "string",
                    "example": "abcdef1234567890"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Optional idempotency key",
            "schema": {
              "type": "string"
            },
            "example": 1234567890
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Unique server generated event ID"
                    },
                    "eventId": {
                      "type": "string",
                      "description": "Tracking ID provided in the request"
                    },
                    "eventType": {
                      "type": "string",
                      "description": "Event type"
                    },
                    "payload": {
                      "type": "object",
                      "description": "Payload"
                    },
                    "channels": {
                      "type": "array",
                      "description": "Channels the webhook was sent to",
                      "items": {
                        "type": "string",
                        "description": "Channel ID"
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "description": "Timestamp the webhook was sent"
                    }
                  },
                  "example": {
                    "id": "abcdef1234567890",
                    "eventId": "abcdef1234567890",
                    "eventType": "user.created",
                    "payload": {
                      "userId": "abcd1234"
                    },
                    "channels": [
                      "ep_a1b2_abcd1234",
                      "grp_a1b2_efgh5678"
                    ],
                    "timestamp": "2021-01-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Usually due to missing parameters, or invalid parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse"
                },
                "example": {
                  "error": "Bad request"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Due to missing or invalid authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse"
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. You do not have permission to access this resource or to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse"
                },
                "example": {
                  "error": "Forbidden"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests. You have exceeded the rate limit. Try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse"
                },
                "example": {
                  "error": "Rate limit exceeded"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. This is a problem with the server that you cannot fix.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse"
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        },
        "x-speakeasy-name-override": "send"
      }
    }
  }
}